requirements.txt
MetaTrader5; python_version < "3.12"
python-dotenv
numpy
